fix bugs and move JW implementation for FermionProduct
